feat: add SD-backed Nova asset management
- upgrade NDI to v2 with 32-bit sector fields for 64MB images\n- extend the CLI with typed remote disk, ROM, soundfont, asset, and device commands\n- stream NovaHost ROM/SID assets from SD instead of embedding generated headers\n- replace ESP async SD HTTP server with a smaller blocking whole-file API\n\nTests:\n- dotnet test e6502.sln -c Release --no-restore\n- make -C e6502.ESP32/novahost compile

+++ e6502.ESP32/novahost/ndi_image.h
@@ -37,12 +37,12 @@ public:
 struct HeaderInfo {
     uint8_t  format_version;
     uint16_t sector_size;
-    uint16_t total_sectors;
+    uint32_t total_sectors;
     char     volume_label[33];           // 32 chars + null
-    uint16_t directory_start_sector;
-    uint16_t directory_sector_count;
-    uint16_t data_start_sector;
-    uint16_t free_sector_count;
+    uint32_t directory_start_sector;
+    uint32_t directory_sector_count;
+    uint32_t data_start_sector;
+    uint32_t free_sector_count;
 };
 
 // File / entry types — match the 6502 wire values.
@@ -68,10 +68,10 @@ struct DirEntry {
     uint8_t  flags;
     uint8_t  file_type;
     uint16_t parent_index;
-    uint16_t start_sector;
+    uint32_t start_sector;
     uint32_t size_bytes;
     char     filename[33];
-    uint16_t sector_count;
+    uint32_t sector_count;
 
     bool is_active()    const { return (flags & FL_ACTIVE)    != 0; }
     bool is_directory() const { return (flags & FL_DIRECTORY) != 0; }
@@ -178,9 +178,9 @@ private:
     // Directory helpers.
     int  dir_find_free_slot();
     void dir_write_entry(int slot, uint8_t flags, uint8_t type,
-                         uint16_t parent, uint16_t start_sector,
+                         uint16_t parent, uint32_t start_sector,
                          uint32_t size, const char* name,
-                         uint16_t sector_count);
+                         uint32_t sector_count);
     void dir_read_entry(int slot, DirEntry& out);
     void dir_clear_slot(int slot);
 };
